server: move the gzip response writer out of server.go
Release note: None
