+++ pkg/server/BUILD.bazel
@@ -21,6 +21,7 @@ go_library(
         "doc.go",
         "drain.go",
         "grpc_server.go",
+        "gzip_response_writer.go",
         "index_usage_stats.go",
         "init.go",
         "init_handshake.go",
