+++ pkg/server/server.go
@@ -12,7 +12,6 @@ package server
 
 import (
 	"bytes"
-	"compress/gzip"
 	"context"
 	"crypto/tls"
 	"encoding/gob"
@@ -117,9 +116,6 @@ import (
 )
 
 var (
-	// Allocation pool for gzipResponseWriters.
-	gzipResponseWriterPool sync.Pool
-
 	forwardClockJumpCheckEnabled = settings.RegisterBoolSetting(
 		settings.TenantWritable,
 		"server.clock.forward_jump_check_enabled",
@@ -2894,61 +2890,6 @@ func (s *Server) PGServer() *pgwire.Server {
 func (s *Server) StartDiagnostics(ctx context.Context) {
 	s.updates.PeriodicallyCheckForUpdates(ctx, s.stopper)
 	s.sqlServer.StartDiagnostics(ctx)
-}
-
-// TODO(benesch): Use https://github.com/NYTimes/gziphandler instead.
-// gzipResponseWriter reinvents the wheel and is not as robust.
-type gzipResponseWriter struct {
-	gz gzip.Writer
-	http.ResponseWriter
-}
-
-func newGzipResponseWriter(rw http.ResponseWriter) *gzipResponseWriter {
-	var w *gzipResponseWriter
-	if wI := gzipResponseWriterPool.Get(); wI == nil {
-		w = new(gzipResponseWriter)
-	} else {
-		w = wI.(*gzipResponseWriter)
-	}
-	w.Reset(rw)
-	return w
-}
-
-func (w *gzipResponseWriter) Reset(rw http.ResponseWriter) {
-	w.gz.Reset(rw)
-	w.ResponseWriter = rw
-}
-
-func (w *gzipResponseWriter) Write(b []byte) (int, error) {
-	// The underlying http.ResponseWriter can't sniff gzipped data properly, so we
-	// do our own sniffing on the uncompressed data.
-	if w.Header().Get("Content-Type") == "" {
-		w.Header().Set("Content-Type", http.DetectContentType(b))
-	}
-	return w.gz.Write(b)
-}
-
-// Flush implements http.Flusher as required by grpc-gateway for clients
-// which access streaming endpoints (as exercised by the acceptance tests
-// at time of writing).
-func (w *gzipResponseWriter) Flush() {
-	// If Flush returns an error, we'll see it on the next call to Write or
-	// Close as well, so we can ignore it here.
-	if err := w.gz.Flush(); err == nil {
-		// Flush the wrapped ResponseWriter as well, if possible.
-		if f, ok := w.ResponseWriter.(http.Flusher); ok {
-			f.Flush()
-		}
-	}
-}
-
-// Close implements the io.Closer interface. It is not safe to use the
-// writer after calling Close.
-func (w *gzipResponseWriter) Close() error {
-	err := w.gz.Close()
-	w.Reset(nil) // release ResponseWriter reference.
-	gzipResponseWriterPool.Put(w)
-	return err
 }
 
 func init() {
